* commit 'c286c6ced31564def0fca7e9f5df63104affaa5b':
  fix packing (#4113)
  fix enable_cache (#4109)
  fix requirements (#4108)
  [megatron] Update long text shell (#4106)
  support max_epochs (#4102)

+++ swift/llm/dataset/utils.py
@@ -4,11 +4,12 @@ import time
 from typing import Any, Callable, Dict, Optional, Union
 
 import numpy as np
+import torch.distributed as dist
 from datasets import Dataset as HfDataset
 from torch.utils.data import Dataset, IterableDataset
 from tqdm import tqdm
 
-from swift.utils import get_logger
+from swift.utils import get_logger, is_dist, is_master
 from ..template import MaxLengthError
 from .preprocessor import RowPreprocessor
 
@@ -152,19 +153,28 @@ class PackingDataset(BasePackingDataset, Dataset):
 
     def __init__(self, template, dataset, num_proc: int = 1, *, packing_interval: int = 128, strict: bool = False):
         super().__init__(template, dataset, num_proc, packing_interval=packing_interval, strict=strict)
-        self.prog_bar = tqdm(total=len(dataset), dynamic_ncols=True, desc='Packing')
+        self.prog_bar = tqdm(total=len(dataset), dynamic_ncols=True, desc=f'Packing (num_proc={num_proc})')
         self._queue = mp.Queue()
         self._terminated_workers = 0
-        for i in range(self.num_proc):
-            shard_dataset = self.dataset.shard(self.num_proc, i)
-            worker = mp.Process(target=self._producer, args=(shard_dataset, ), daemon=True)
-            worker.start()
-            self.workers.append(worker)
+        if is_master():
+            for i in range(self.num_proc):
+                shard_dataset = self.dataset.shard(self.num_proc, i)
+                worker = mp.Process(target=self._producer, args=(shard_dataset, ), daemon=True)
+                worker.start()
+                self.workers.append(worker)
 
-        self.packed_dataset = self.get_packed_dataset()
-        self.prog_bar.close()
-        for worker in self.workers:
-            worker.terminate()
+            self.packed_dataset = self.get_packed_dataset()
+            self.prog_bar.close()
+            for worker in self.workers:
+                worker.terminate()
+            if is_dist():
+                obj_list = [self.packed_dataset]
+                dist.broadcast_object_list(obj_list)
+                self.packed_dataset = obj_list[0]
+        elif is_dist():
+            obj_list = [None]
+            dist.broadcast_object_list(obj_list)
+            self.packed_dataset = obj_list[0]
 
     def fetch_packing_data(self, res: Optional[list] = None):
         res = res or []
@@ -223,10 +233,11 @@ class IterablePackingDataset(BasePackingDataset, IterableDataset):
         self._out_queue = mp.Queue()
         self.workers = []
         self.cyclic = cyclic
-        for _ in range(self.num_proc):
-            worker = mp.Process(target=self._processor, daemon=True)
-            worker.start()
-            self.workers.append(worker)
+        if is_master():
+            for _ in range(self.num_proc):
+                worker = mp.Process(target=self._processor, daemon=True)
+                worker.start()
+                self.workers.append(worker)
 
     def _processor(self):
         while True:
@@ -264,11 +275,13 @@ class IterablePackingDataset(BasePackingDataset, IterableDataset):
                 yield x
 
     def __iter__(self):
+        try:
+            next(iter(self.dataset))
+        except StopIteration:
+            return
+
+        assert len(self.workers) > 0, f'self.workers: {self.workers}'
         if self.cyclic:
-            try:
-                next(iter(self.dataset))
-            except StopIteration:
-                return
             iterator = self.cyclic_iter(self.dataset)
         else:
             iterator = iter(self.dataset)
